NODE-102, fix: use correct uint type

+++ pallets/btc-registration-pool/src/pallet/impls.rs
@@ -41,7 +41,7 @@ impl<T: Config> PoolManager<T::AccountId> for Pallet<T> {
 		T::BitcoinNetwork::get()
 	}
 
-	fn get_bitcoin_chain_id() -> u64 {
+	fn get_bitcoin_chain_id() -> u32 {
 		T::BitcoinChainId::get()
 	}
 }
